+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.0.4
+- [Added] Flutter Counter example
+- [Added] getModel<T> helper function
+- [Changed] Updated ViewModelBuilder Live Template to use getModel<T>
+- **Date**: 8-14-2023
+
 ## 0.0.3
 - [Changed] Minor updates to readme
 - **Date**: 8-13-2023